requirements.txt
python-dotenv==1.0.0
openai==1.12.0
requests==2.31.0
yt-dlp==2023.12.30
moviepy==1.0.3
transformers==4.38.2
torch==2.2.1
tf-keras==2.15.0 